Major changes, update SDK, clean paths and more
