+++ app/sign-up/address-register.tsx
@@ -1,17 +1,14 @@
 import React from "react";
-import { StyleSheet, TouchableOpacity, View } from "react-native";
-import { useRouter } from "expo-router";
-import Icon from "react-native-vector-icons/MaterialIcons";
-import ProgressBarContainer from "@/components/Visuals/ProgressBarContainer";
+import ProgressBarContainer from "@/components/visuals/ProgressBarContainer";
 import LayoutRegister from "@/components/layouts/register/LayoutRegister";
 import Title from "@/components/layouts/Title";
 import AddressRegistrationForm from "@/components/forms/register/AddressRegistrationForm";
 
 import ScrollLayout from "@/components/layouts/ScrollLayout";
 import HeaderRegister from "@/components/layouts/register/HeaderRegister";
-import { AbsoluteBackButton } from "@/components/Buttons/navigation/AbsoluteBackButton";
+import { AbsoluteBackButton } from "@/components/buttons/navigation/AbsoluteBackButton";
 
-export default function () {
+export default function SecondRegister() {
   const eventRedirect = async () => {
     console.log("Redirect to Register 3...");
   };
@@ -28,7 +25,7 @@ export default function () {
       <LayoutRegister>
         <ScrollLayout>
           <AddressRegistrationForm
-            redirect={"/sign-up/register3"}
+            redirect={"/sign-up/select-register"}
             onSubmit={eventRedirect}
           />
         </ScrollLayout>
