--- app/sign-up/termsAndConditions.tsx
@@ -1,25 +1,0 @@
-import React from "react";
-import { StyleSheet, Text, View } from "react-native";
-
-const termsAndConditions = () => {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.text}>This is the terms and conditions page.</Text>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "#fff",
-  },
-  text: {
-    fontSize: 20,
-    fontWeight: "bold",
-  },
-});
-
-export default termsAndConditions;
