+++ components/cards/CardVehicleInfo.tsx
