+++ components/forms/register/AddressRegistrationForm.tsx
@@ -2,10 +2,10 @@ import React, { useState } from "react";
 import { router } from "expo-router";
 import { StyleSheet } from "react-native";
 
-import InputLabel from "@/components/Inputs/InputLabel";
-import InputPassword from "@/components/Inputs/InputPassword";
-import ButtonSubmit from "@/components/Buttons/ButtonSubmit";
+import InputLabel from "@/components/inputs/InputLabel";
+import InputPassword from "@/components/inputs/InputPassword";
 import { RegistrationFormProps } from "@/components/forms/register/RegisterForm";
+import { InputButton } from "@/components/buttons/InputButton";
 
 export default function AddressRegistrationForm({
   onSubmit,
@@ -47,61 +47,17 @@ export default function AddressRegistrationForm({
         label="Confirmar Contraseña"
         onChangeText={setConfirmPassword}
       />
-      <ButtonSubmit
-        title="Siguiente"
+      <InputButton
+        label={"Siguente"}
+        typeButton={"submit"}
         onPress={onPressButton}
-        style={styles.submitButton}
       />
     </>
   );
 }
 
 const styles = StyleSheet.create({
-  flexContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    width: "100%",
-    padding: 0,
-  },
-  backgroundContainer: {
-    flex: 1,
-    backgroundColor: "#0f2422",
-  },
-  headerContainer: {
-    width: "100%",
-    alignItems: "center",
-    paddingTop: 60,
-    paddingBottom: 20,
-  },
-  progressBar: {
-    marginTop: 15,
-    width: "60%",
-  },
-  backButton: {
-    position: "absolute",
-    top: 43,
-    left: 20,
-  },
-  formContainer: {
-    flex: 1,
-    backgroundColor: "#fff",
-    borderTopLeftRadius: 40,
-    paddingTop: 30,
-    alignItems: "center",
-    justifyContent: "flex-start",
-    paddingHorizontal: 20,
-  },
-  title: {
-    fontSize: 43.2,
-    fontWeight: "bold",
-    color: "#fff",
-    textAlign: "center",
-  },
   userInputMargin: {
     marginTop: 10,
-  },
-  submitButton: {
-    marginTop: 24,
   },
 });
