+++ components/sheets/BottomSheet.jsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { View } from "react-native";
 import styles from "@/styles/Sheets";
 
-const BottomSheet = () => {
+const BottomSheet = (props) => {
   return (
     <View style={styles.bottomSheet}>
       <View style={styles.bottomSheetGrabHandle} />
