+++ components/visuals/ProgressBarContainer.tsx
