+++ stories/CustomRadioButtonGroup.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import React from "react";
 import { View } from "react-native";
-import CustomRadioButtonGroup from "@/components/Cards/CustomRadioButtonGroup";
+import CustomRadioButtonGroup from "@/components/cards/CustomRadioButtonGroup";
 import * as Crypto from "expo-crypto";
 
 const DATA = [
