+++ stories/Map.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import React from "react";
 import { View } from "react-native";
-import Map from "@/components/Cards/Map";
+import Map from "@/components/cards/Map";
 
 const meta = {
   title: "Map",
